drivers/imx6-enet: rename files
JIRA: RTOS-507
